Optimisations & fix for issue #20 using int buffers instead of short buffers

+++ src/rajawali/IObject3D.java
@@ -11,7 +11,7 @@ public interface IObject3D {
 	public void render(Camera camera, float[] projMatrix, float[] vMatrix, ColorPickerInfo pickerInfo);
 	public void render(Camera camera, float[] projMatrix, float[] vMatrix, float[] parentMatrix, ColorPickerInfo pickerInfo);
 	public void addTexture(TextureInfo textureInfo);
-	public void setData(float[] vertices, float[] normals, float[] textureCoords, float[] colors, short[] indices);
+	public void setData(float[] vertices, float[] normals, float[] textureCoords, float[] colors, int[] indices);
 
 	public void setScreenCoordinates(float x, float y, int viewportWidth, int viewportHeight, float eyeZ);
